Clean perception common (#2164)
* clear compiling warning in hough_transfer.cc

fixed some warnings in camera_homography.cc

fixed some warnings in conditional_clustering_test.cc

clearn warnings in basic.h

clear warnings in conditional_clustering.h

clear warnings in convex_hull_2d_test.cc

fixed a warning in convex_hull_2d.h

clean warnings in downsampling_test.cc

clean warnings in downsampling.h

clean warnings in i_plane.h

clean warnings in common.h

clean warnings in io_util.cc

fix bug induced when cleaning

* small modify of a fix

* address xiangquans comment

* address xiangquans comment

+++ modules/perception/common/geometry/camera_homography.cc
@@ -53,7 +53,9 @@ bool IsCamerasFieldOverlap(const base::PinholeCameraModel& from_camera,
   size_t to_height = to_camera.get_height();
   Eigen::Matrix3d f2t_rotation = extrinsic.topLeftCorner(3, 3);
   Eigen::Matrix<double, 4, 2> points;
-  points << 0.0, 0.0, f_width, 0.0, f_width, f_height, 0.0, f_height;
+  points << 0.0, 0.0, static_cast<double>(f_width), 0.0,
+            static_cast<double>(f_width), static_cast<double>(f_height),
+            0.0, static_cast<double>(f_height);
   Eigen::Vector2d pt_min(std::numeric_limits<double>::max(),
                          std::numeric_limits<double>::max());
   Eigen::Vector2d pt_max(-std::numeric_limits<double>::max(),
